requirements.txt
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
jsonfield==3.1.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycosat==0.6.3
Pygments==2.10.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.3.0
django_heroku
